C0000IPCDM-149 Improve collectionview delegate and datasource in Base
Why:
   - The developer wants the delegate and datasource in the presenter.
How:
   - Created a collectionview base that implements the delegates.
   - Created a cell base to change highlight color.

+++ CoVi/Common/Protocols/TableView/CoViTableViewDelegate.swift
@@ -6,14 +6,12 @@
 //  Copyright © 2020 Babel. All rights reserved.
 //
 
-import UIKit
-
 /**
  CoVi base table view delegate.
  */
 @objc public protocol CoViTableViewDelegate: class {
-    @objc optional func heightForRowAt(_ tableViewTag: Int, _ indexPath: IndexPath) -> CGFloat
-    @objc optional func estimatedHeightForRowAt(_ tableViewTag: Int, _ indexPath: IndexPath) -> CGFloat
+    @objc optional func heightForRowAt(_ tableViewTag: Int, _ indexPath: IndexPath) -> Float
+    @objc optional func estimatedHeightForRowAt(_ tableViewTag: Int, _ indexPath: IndexPath) -> Float
     @objc optional func didSelectRowAt(_ tableViewTag: Int, _ indexPath: IndexPath)
     @objc optional func didDeselectRowAt(_ tableViewTag: Int, _ indexPath: IndexPath)
     @objc optional func willDisplay(_ tableViewTag: Int, _ indexPath: IndexPath)
